day80/requirements.txt
pandas
matplotlib
PyQt6
plotly-express

